+++ src/onegov.event/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Allows to filter occurrences by locations.
+  [msom]
+
 1.0.0 (2018-12-07)
 ~~~~~~~~~~~~~~~~~~~